refactor(napi): narrow tokio runtime changes
