+++ crates/rspack_napi/Cargo.toml
@@ -11,8 +11,8 @@ version.workspace = true
 napi         = { workspace = true, features = ["serde-json", "anyhow", "napi9", "compat-mode"] }
 oneshot      = { workspace = true }
 rspack_error = { workspace = true }
-rspack_tasks = { workspace = true }
 serde_json   = { workspace = true }
+tokio        = { workspace = true, features = ["rt", "rt-multi-thread", "sync", "time"] }
 
 [lints]
 workspace = true
