+++ xtask/benchmark/benches/groups/bundle.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use criterion::{Criterion, SamplingMode};
+use criterion::Criterion;
 use rspack_tasks::{CompilerContext, within_compiler_context, within_compiler_context_sync};
 
 use crate::groups::{
@@ -14,13 +14,6 @@ pub mod threejs;
 pub mod util;
 
 pub(crate) fn bundle_benchmark_case(c: &mut Criterion, target_id: &str) {
-  // CodSpeed simulation instruments every instruction, making full bundle cases too large for
-  // the PR job timeout. Walltime and regular benchmark runs still exercise these cases.
-  #[cfg(codspeed)]
-  if rspack_benchmark::is_simulation_benchmark() {
-    return;
-  }
-
   let projects: Vec<(&'static str, CompilerBuilderGenerator)> = vec![
     ("basic-react", Arc::new(basic_react::compiler)),
     ("misc", Arc::new(misc::compiler)),
@@ -34,9 +27,6 @@ pub(crate) fn bundle_benchmark_case(c: &mut Criterion, target_id: &str) {
   // Codspeed can only handle to up to 500 threads by default
   let rt = rspack_benchmark::build_tokio_rt();
   let mut group = c.benchmark_group("bundle");
-  if rspack_benchmark::is_simulation_benchmark() {
-    group.sample_size(10).sampling_mode(SamplingMode::Flat);
-  }
 
   group.bench_function(format!("bundle@{id}"), |b| {
     b.iter_batched(
